requirements.txt
black==24.3.0
Django==4.1
djangorestframework==3.14.0
psycopg2==2.9.6
